requirements.txt
langchain
langchain-groq
faiss-cpu
sentence-transformers
pandas
flask
python-dotenv
langchain_community
langchain-huggingface
gunicorn

gunicorn
python-dotenv
spacy==3.7.5

lxml
beautifulsoup4
PyPDF2

soundfile
pydub
kokoro